Added stair functionality
